Modified baseline calculation.
The previous baseline calculation did not consider the number of
iterations used within a single test.  All demos used the same number by
comparison, so this bug did not show up.  Now, the baseline comparison
assumes that the body of each benchmark or baseline function body has
equal weight within a specific benchmark case.  This seems more
intuitive and should provide more accurate baseline results.

+++ include/celero/BenchmarkInfo.h
@@ -89,12 +89,12 @@ namespace celero
 			///
 			///
 			///
-			void setBaselineUnit(const uint64_t x);
+			void setBaselineUnit(const double x);
 			
 			///
 			///
 			///
-			uint64_t getBaselineUnit() const;
+			double getBaselineUnit() const;
 
 			///
 			/// \brief	Sets the test's best run time.
